Cleanup (for proper API signature regarding SparkDatasource with DataFrame DataAsset).
