+++ great_expectations/datasource/fluent/schemas/SparkDatasource.json
@@ -1,13 +1,13 @@
 {
-    "title": "SparkDataframeDatasource",
+    "title": "SparkDatasource",
     "description": "Base model for most fluent datasource related pydantic models.\n\nAdds yaml dumping and parsing methods.\n\nExtra fields are not allowed.\n\nSerialization methods default to `exclude_unset = True` to prevent serializing\nconfigs full of mostly unset default values.\nAlso prevents passing along unset kwargs to BatchSpec.\nhttps://docs.pydantic.dev/usage/exporting_models/",
     "type": "object",
     "properties": {
         "type": {
             "title": "Type",
-            "default": "spark_dataframe",
+            "default": "spark",
             "enum": [
-                "spark_dataframe"
+                "spark"
             ],
             "type": "string"
         },
